+++ bot.py
@@ -1,10 +1,11 @@
 import os
 from sched import scheduler
 from discord.ext import commands
-import discord
+from discord import option, permissions
 from dotenv import load_dotenv
 import zu_api
 import sql
+import modals
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 from datetime import datetime
 
@@ -12,7 +13,7 @@ from datetime import datetime
 load_dotenv(dotenv_path="config")
 
 #Init the discord bot and set command prefix
-bot = commands.Bot(command_prefix="!")
+bot = commands.Bot()
 
 #If not exist: creation of the db
 sql.CreateDB()
@@ -20,67 +21,51 @@ sql.CreateDB()
 ################
 # BOT COMMANDS #
 ################
-
-#Test command used to ckeck if bot is alive
-@bot.command(name="ping")
-async def ping(ctx, end=None):
-    if end is None:
-        await ctx.channel.send("pong")
-    else:
-        await ctx.channel.send("Tu as entré trop d'arguments")
+@bot.slash_command(
+    name="ping",
+    description="Vérifier si le bot est actif."
+)
+async def ping(ctx):
+    await ctx.respond("pong")
 
 #Command used to add user to subscriber list
-@bot.command(name="sub")
-async def sub(ctx, end=None):
+@bot.slash_command(
+    name="sub",
+    description="S'abonner aux alertes"
+)
+async def sub(ctx):
     #Command only available in sub-unsub channel
-    if int(ctx.channel.id) != int(os.getenv("SUB_UNSUB_CHANNEL")):
-        print (ctx.channel.id)
-        return
 
-    if end is None:
-        if zu_api.CheckUser(str(ctx.message.author.name), str(ctx.message.author.discriminator)): #Check if user is a ZUnivers player
-            if sql.AddUser(str(ctx.message.author.id), str(ctx.message.author.name), str(ctx.message.author.discriminator)):
-                sub_role = discord.utils.get(ctx.message.guild.roles, name="sub")
-                not_sub_role = discord.utils.get(ctx.message.guild.roles, name="not-sub")
-
-                await ctx.message.author.add_roles(sub_role)
-                await ctx.message.author.remove_roles(not_sub_role)
-                await ctx.channel.send(ctx.message.author.mention + " : Tu rejoins les élus en t'abonnant aux alertes !")
-            else:
-                await ctx.channel.send(ctx.message.author.mention + " : Tu es déjà abonné aux alertes.")
+    if zu_api.CheckUser(str(ctx.author.name), str(ctx.author.discriminator)): #Check if user is a ZUnivers player
+        if sql.AddUser(str(ctx.author.id), str(ctx.author.name), str(ctx.author.discriminator)):
+            await ctx.respond(ctx.author.mention + " : Tu rejoins les élus en t'abonnant aux alertes !")
         else:
-            await ctx.channel.send(ctx.message.author.mention + " : Tu n'es pas encore membre ZUnivers. Commence par faire un !journa dans le serveur officiel.")
+            await ctx.respond(ctx.author.mention + " : Tu es déjà abonné aux alertes.")
     else:
-        await ctx.channel.send("Tu as entré trop d'arguments")
+        await ctx.respond(ctx.author.mention + " : Tu n'es pas encore membre ZUnivers. Commence par faire un !journa dans le serveur officiel.")
 
 #Command used to unscribe to the alerts
-@bot.command(name="unsub")
-async def unsub(ctx, end=None):
+@bot.slash_command(
+    name="unsub",
+    description="Se désabonner des alertes"
+)
+async def unsub(ctx):
     #Command only available in sub-unsub channel
-    if int(ctx.channel.id) != int(os.getenv("SUB_UNSUB_CHANNEL")):
-        print (ctx.channel.id)
-        return
 
-    if end is None:
-        if sql.DelUser(str(ctx.message.author.id)):
-            sub_role = discord.utils.get(ctx.message.guild.roles, name="sub")
-            not_sub_role = discord.utils.get(ctx.message.guild.roles, name="not-sub")
-
-            await ctx.message.author.add_roles(not_sub_role)
-            await ctx.message.author.remove_roles(sub_role)
-            await ctx.channel.send(ctx.message.author.mention + " : Tu es désabonné, tu ne peux plus compter que sur toi-même...")
-        else:
-            await ctx.channel.send(ctx.message.author.mention + " : Il faudrait déjà être abonné avant de se désabonner...")
+    if sql.DelUser(str(ctx.author.id)):
+        await ctx.respond(ctx.author.mention + " : Tu es désabonné, tu ne peux plus compter que sur toi-même...")
     else:
-        await ctx.channel.send("Tu as entré trop d'arguments")
+        await ctx.respond(ctx.author.mention + " : Il faudrait déjà être abonné avant de se désabonner...")
 
 ############################
 # ADMIN COMMANDS FOR DEBUG #
 ############################
 
 #Show the list of suscribed users
-@bot.command(name="list")
-@commands.has_role('admin')
+@bot.slash_command(
+    name="list",
+    description="Affiche la liste de tous les utilisateurs."
+)
 async def list(ctx):
     list_users = sql.GetUsers()
 
@@ -88,38 +73,63 @@ async def list(ctx):
         for user in list_users:
             loot = zu_api.CheckActivity(str(user[1]), str(user[2]))
             tower = zu_api.CheckAs(str(user[1]), str(user[2]))
-            await ctx.channel.send(str(user[1]) + "#" + str(user[2]) + " | Loot =  " + str(loot) + " | As =  " + str(tower))
+            await ctx.respond(str(user[1]) + "#" + str(user[2]) + " | Loot =  " + str(loot) + " | As =  " + str(tower))
     else:
-        await ctx.channel.send("Il n'y a pas d'utilisateurs enregistrés.")
+        await ctx.respond("Il n'y a pas d'utilisateurs enregistrés.")
 
 #Test private message
-@bot.command(name="mp")
-@commands.has_role('admin')
+@bot.slash_command(
+    name="mp",
+    description="Test d'envoi des messages privés."
+)
 async def mp(ctx):
     user = await bot.fetch_user(594195299967434773)
     await user.send('Hello World')
+    await ctx.respond("Message envoyé.")
 
 #Manually send a reminder
-@bot.command(name="check")
-@commands.has_role('admin')
+@bot.slash_command(
+    name="check",
+    description="Enovoyer tous les rappels manuellement."
+)
 async def check(ctx):
     await SendJournaReminder()
     await SendAsReminder()
+    await ctx.respond("Rappels envoyés.")
 
 #Send message from bot to a specific channel
-@bot.command(name="send")
-@commands.has_role('admin')
-async def send(ctx, channelID, message, end=None):
-    if end is None:
-        channel = bot.get_channel(int(channelID))
-        await channel.send(str(message))
-    else:
-        await ctx.channel.send("Tu as entré trop d'arguments")
+@bot.slash_command(
+    name="send",
+    description="Envoyer un message depuis le bot"
+)
+@option(
+    "channel",
+    description="Channel où envoyer le message",
+    choices=["General", "Sub-Unsub", "Admin"]
+)
+@option(
+    "tags",
+    description="Tags à jouter au début du message",
+    required=False
+)
+async def send(ctx, channel, tags):
+    if channel == "General":
+        channel_id = bot.get_channel(int(os.getenv("GENERAL_CHANNEL")))
+    if channel == "Admin":
+        channel_id = bot.get_channel(int(os.getenv("ADMIN_CHANNEL")))
+    if channel == "Sub-Unsub":
+        channel_id = bot.get_channel(int(os.getenv("SUB_UNSUB_CHANNEL")))
 
-@bot.command(name="time")
-@commands.has_role('admin')
+    modal = modals.Send_Modal(title="Envoyer message", channel_id=channel_id, tags=str(tags))
+    await ctx.send_modal(modal)
+
+#Print local time
+@bot.slash_command(
+    name="time",
+    description="Afficher l'heure du serveur."
+)
 async def time(ctx):
-    await ctx.channel.send(str(datetime.now()))
+    await ctx.respond(str(datetime.now()))
 
 ####################
 # COMMON FUNCTIONS #
@@ -174,4 +184,4 @@ scheduler.start()
 # START BOT #
 #############
 
-bot.run(os.getenv("TOKEN"))
+bot.run(token=os.getenv("TOKEN"))